requirements.txt
edgartools==3.0.3
flask==3.0.3
waitress==3.0.0
PyYAML==6.0.2